% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataconnector.R
\docType{package}
\name{dataconnector}
\alias{dataconnector}
\alias{dataconnector-package}
\title{\pkg{dataconnector} allows users to read and write files from HDFS.}
\description{
\pkg{dataconnector} is extensible and new file formats and filesystems can be easily added. So far it supports:
}
\details{
Formats:
\itemize{
  \item CSV
  \item ORC
}
Filesystems:
\itemize{
  \item HDFS
  \item Local FS
}
}

