% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdfsConnector.R
\docType{package}
\name{hdfsconnector}
\alias{hdfsconnector}
\alias{hdfsconnector-package}
\title{\pkg{hdfsconnector} allows users to read and write files from HDFS.}
\description{
\pkg{hdfsconnector} is extensible and new file formats and filesystems can be easily added. So far it supports:
}
\details{
Formats:
\itemize{
  \item CSV
  \item ORC
}
Filesystems:
\itemize{
  \item HDFS
  \item Local FS
}
}

